New Project
I copied over most of the things from the vulkan-coffee demo,
and trimmed a few things too. Gone is EASTL, because I don't
need that kind of complexity in my life. I shuffled some things
around in the move, but everything is here (or will be in a couple
of commits). This also acts as an exercise in architecting a simple
Vulkan app. We'll see.

Hopefully we can get things working with Katy models soon! :D